api/requirements.txt
fastapi>=0.85.0,<0.86.0
pydantic[email]>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
psycopg2>=2.9.0,<2.10.0
sqlalchemy>=1.4.0,<1.5.0
python-dotenv>=0.21.0,<0.22.0
geoalchemy2>0.12.0,<0.13.0
shapely>=1.8.0,<1.9.0
python-multipart
python-jose[cryptography]
passlib[bcrypt]

# dev
pytest>=7.1.0,<7.2.0
requests>=2.28.0,<2.29.0